rename page to `manage-content-page`
