+++ kolibri/plugins/management/assets/src/vue/manage-content-page/index.vue
