+++ kolibri/plugins/management/assets/src/vue/manage-content-page/modal.vue
